django_history_service/requirements.txt
django==4.2
djangorestframework==3.15.1
django-cors-headers==4.3.1